thread_pract10.java
This code demonstrates how to create a custom thread in Java. The class MyThread extends the built-in Thread class, providing a run() method that gets executed when the thread starts.  The run() method simply prints the name of the current thread. In the main method, a MyThread object is created with the name "Ishan" and then started, causing "Ishan" to be printed to the console.