thread_sync3.java
User Input: The main change here is the introduction of a Scanner object (sc) in the Main1 class. This allows users to directly type in how many seats they want to book during the program's execution.
Waiting for Turn: The t1.join() line is interesting. This makes the main program wait until thread t1 (the first customer booking) is finished before asking the second customer for input.