thread_sync5.java
It creates the child thread and tells it to start working.
It announces "Main Thread is Waiting". Then, it takes a break using the wait() command. It's basically waiting for the child thread to finish its job.
When the child thread sends the notification, the main thread wakes up! "Main Thread gets Notification.."
Finally, it gets the result of the calculation from the child thread and proudly displays it.






synchronized:  This word is like a special room. Only one thread can be inside the room at a time. It makes sure things happen in order and prevents conflicts.

wait() :   This tells a thread to take a nap inside the synchronized room. It will sleep until someone else wakes it up.

notify(): This is how one thread wakes up another thread that's sleeping with wait().